requirements.txt
boto3==1.7.28
botocore==1.10.28
docutils==0.14
jmespath==0.9.3
mysqlclient==1.3.12
python-dateutil==2.7.3
PyYAML==3.12
s3transfer==0.1.13
six==1.11.0
SQLAlchemy==1.2.7
xmltodict==0.11.0
